Added further improvements of namespaces
