+++ Gemfile
@@ -1,0 +1,2 @@
+gem "ruby2d"
+gem "aasm"
